// frontend/lib/api/generated/model/characterCreateImportanceLevel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * AI小说创作系统
 * 作者主导的长篇小说创作平台：常驻创作 Agent、固定生成工作流与可编排的复合任务
 * OpenAPI spec version: 0.1.0
 */

/**
 * 重要性级别
 */
export type CharacterCreateImportanceLevel = typeof CharacterCreateImportanceLevel[keyof typeof CharacterCreateImportanceLevel];


export const CharacterCreateImportanceLevel = {
  main: 'main',
  secondary: 'secondary',
  minor: 'minor',
} as const;