// frontend/lib/api/generated/model/characterUpdateImportanceLevel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * AI小说创作系统
 * 作者主导的长篇小说创作平台：常驻创作 Agent、固定生成工作流与可编排的复合任务
 * OpenAPI spec version: 0.1.0
 */

export type CharacterUpdateImportanceLevel = typeof CharacterUpdateImportanceLevel[keyof typeof CharacterUpdateImportanceLevel] | null;


export const CharacterUpdateImportanceLevel = {
  main: 'main',
  secondary: 'secondary',
  minor: 'minor',
} as const;